PW-5490 - Move all web api endpoints, except stateData to the API directory
